Add auto-tuning options
--size option sets the size (in bytes) the of the benchmark working set.
Suffixes kK, mM, and gG are allowed. The default is the L1 cache size as
reported by hwloc. To be used with --tune or --auto. Has no effect on its own.

--time option sets the benchmark runtime (per iteration) in seconds. The default
is 20 seconds. To be used with --tune or --auto. Has no effect on its own.

--tune suggests benchmark size and rounds parameters. First, the benchmark size
parameter is determined such that the benchmark uses approximately 90% of the
amount specified via --size. Second, the rounds parameter is calculated such
that one benchmark iteration takes approximaltely the time specified by the
--time parameter. The resulting parameters are printed to stderr.

--auto calculates the parameters, same as --tune, but also runs the benchmarks
with the calculated options. For reproducibility, --tune is suggested over
--auto.

If no benchmark is specified with --benchmarks, all benchmarks are selected.