Bugfix x86_64 PMU backend selection logic
Of course we can't select the PMU backend for Linux or McKernel on compile
time, because compiling for McKernel means compiling for Linux and vice versa.
Instead select the backends independently at compile time. This get's messy
because the test whether or not we're running on McKernel is a run-time test.